pyproject.toml
[tool.poetry]
name = "inky-dash"
version = "2.3.2"
description = "Inky dash is an interface for Inky displays, a line of e-paper displays for the Raspberry Pi, that can be accessed from a browser via a local web server."
authors = ["Robert <robert@rrayns.co.uk>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
altgraph = "0.17.4"
annotated-types = "0.6.0"
blinker = "1.6.2"
click = "8.1.7"
dependency-injector = "4.43.0"
exceptiongroup = "1.2.1"
flask = "2.3.3"
flask-cors = "4.0.1"
gunicorn = "21.2.0"
iniconfig = "2.0.0"
inky = "1.5.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.4"
markupsafe = "2.1.3"
numpy = "1.26.4"
packaging = "23.2"
pillow = "10.3.0"
pluggy = "1.5.0"
pydantic = "2.6.3"
pydantic-core = "2.16.3"
pyinstaller = "6.1.0"
pyinstaller-hooks-contrib = "2023.10"
pytest = "8.2.0"
six = "1.16.0"
smbus2 = "0.4.3"
spidev = "3.6"
tomli = "2.0.1"
typing-extensions = "4.10.0"
werkzeug = "2.3.7"
rpi-gpio = "^0.7.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
